Replace float with real_t, default Vectors, other misc C# improvements
Replace float with real_t in most files, defined at the top of each file via using. Objects such as Vector3 now accept doubles as inputs, and convert to real_t internally. I've added default Vectors such as Vector3.Zero. Other misc C# improvements such as Mathf.RoundToInt(). Color continues to use float only because high precision is not needed for 8-bit color math and to keep things simple. Everything seems to compile and work fine, but testing is requested, as this is the first time I've ever contributed to Godot.

(cherry picked from commit ff97c97c930f80209480c630622b1b64372bd65a)

+++ modules/mono/glue/cs_files/Color.cs
@@ -45,8 +45,8 @@ namespace Godot
         {
             get
             {
-                float max = Mathf.Max(r, Mathf.Max(g, b));
-                float min = Mathf.Min(r, Mathf.Min(g, b));
+                float max = (float) Mathf.Max(r, (float) Mathf.Max(g, b));
+                float min = (float) Mathf.Min(r, (float) Mathf.Min(g, b));
 
                 float delta = max - min;
 
@@ -79,8 +79,8 @@ namespace Godot
         {
             get
             {
-                float max = Mathf.Max(r, Mathf.Max(g, b));
-                float min = Mathf.Min(r, Mathf.Min(g, b));
+                float max = (float) Mathf.Max(r, (float) Mathf.Max(g, b));
+                float min = (float) Mathf.Min(r, (float) Mathf.Min(g, b));
 
                 float delta = max - min;
 
@@ -96,7 +96,7 @@ namespace Godot
         {
             get
             {
-                return Mathf.Max(r, Mathf.Max(g, b));
+                return (float) Mathf.Max(r, (float) Mathf.Max(g, b));
             }
             set
             {
@@ -316,7 +316,8 @@ namespace Godot
 
             return txt;
         }
-
+        
+        // Constructors 
         public Color(float r, float g, float b, float a = 1.0f)
         {
             this.r = r;
@@ -375,7 +376,7 @@ namespace Godot
 
         private String _to_hex(float val)
         {
-            int v = (int)Mathf.Clamp(val * 255.0f, 0, 255);
+            int v = (int) Mathf.Clamp(val * 255.0f, 0, 255);
 
             string ret = string.Empty;
